update to the forum comment logic

+++ controllers/API/comment.js
@@ -1,35 +1,36 @@
 var express = require('express');
-var router  = express.Router();
+var router = express.Router();
 var Comment = require('mongoose').model('comment');
 var Forum = require('mongoose').model('forum');
 
 
-router.post('/API/forum/comment',function (req, res) {
+router.post('/API/forum/comment', function (req, res) {
     var newComment = new Comment({
         text: req.body.text,
         author: req.body.author,
         posterID: req.body.posterID,
-        imgURL:req.body.imgURL
+        imgURL: req.body.imgURL,
+        parentThread: req.body.threadID
     });
     newComment.save(function (err, result) {
         if (err) {
             res.status(201);
             res.send(err);
-            console.log("error creating the comment",err);
+            console.log("error creating the comment", err);
         } else {
-            Forum.findById(req.body.threadID,function(err2,thread){
-                if(err2){
+            Forum.findById(req.body.threadID, function (err2, thread) {
+                if (err2) {
                     res.status(201);
                     res.send(err2);
-                    console.log("error finding thread with the threadID in the request",err2);
+                    console.log("error finding thread with the threadID in the request", err2);
                 }
                 else {
                     thread.comments.push(result._id);
-                    thread.save(function(err3){
-                        if(err3){
+                    thread.save(function (err3) {
+                        if (err3) {
                             res.status(201);
                             res.send(err3);
-                            console.log("error updating comment in thread array",err3);
+                            console.log("error updating comment in thread array", err3);
                         }
                         else {
                             res.send(200);
@@ -43,16 +44,40 @@ router.post('/API/forum/comment',function (req, res) {
 });
 
 router.delete('/API/forum/comment', function (req, res) {
-    Comment.findOneAndRemove({
+    Comment.findOne({
         '_id': req.param('id')
-    }, function (err, user) {
-        if (err){
+    }, function (err, comment) {
+        if (err) {
             res.status(201);
             res.send(err);
-            console.log(err)
+            console.log("error finding comment with the id provided", err)
         }
         else {
-            res.send(200);
+            Forum.findOneAndUpdate({
+                    '_id': comment.parentThread
+                },
+                {$pop: {"comments": req.param('id')}},
+                function (err2) {
+                    if (err2) {
+                        res.status(201);
+                        res.send(err2);
+                        console.log("error finding the parent thread of the comment", err2)
+                    }
+                    else {
+                        Comment.findOneAndRemove({
+                            '_id': req.param('id')
+                        }, function (err3) {
+                            if (err3) {
+                                res.status(201);
+                                res.send(err3);
+                                console.log("error deleting comment with specified id", err3)
+                            }
+                            else {
+                                res.send(200);
+                            }
+                        });
+                    }
+                });
         }
     });
 
@@ -67,7 +92,7 @@ router.put('/API/forum/comment', function (req, res) {
             edited: Date.now()
         }
     }, function (err, user) {
-        if (err){
+        if (err) {
             res.status(201);
             res.send(err);
             console.log(err);
@@ -79,9 +104,9 @@ router.put('/API/forum/comment', function (req, res) {
 });
 router.get('/API/forum/comment', function (req, res) {
     Comment
-        .findOne({'_id':req.param('id')})
+        .findOne({'_id': req.param('id')})
         .exec(function (err, comment) {
-            if (err){
+            if (err) {
                 res.status(201);
                 res.send(err);
                 console.log(err);
